Break transform and job configs into mirror programs
