+++ CODEOWNERS
@@ -2,6 +2,7 @@
 .prow.yaml                           @istio/wg-test-and-release-maintainers @istio/operations-maintainers
 /authentikos/                        @istio/wg-test-and-release-maintainers @istio/operations-maintainers
 /prow/                               @istio/operations-maintainers
+/prow/generate-transform-jobs/       @istio/wg-test-and-release-maintainers
 /prow/config/                        @istio/wg-test-and-release-maintainers
 /prow/config/jobs/test-infra.yaml    @istio/wg-test-and-release-maintainers @istio/operations-maintainers
 /prow/cluster/jobs/                  @istio/wg-test-and-release-maintainers
