+++ prow/cluster/jobs/istio-private/istio.io/istio-private.istio.io.release-1.10.gen.yaml
@@ -15,7 +15,7 @@ postsubmits:
       - command:
         - make
         - lint
-        image: gcr.io/istio-testing/build-tools:release-1.10-2021-06-17T16-10-50
+        image: gcr.io/istio-testing/build-tools:release-1.10-2021-07-06T17-17-03
         name: ""
         resources:
           limits:
@@ -51,7 +51,7 @@ postsubmits:
       - command:
         - make
         - gen-check
-        image: gcr.io/istio-testing/build-tools:release-1.10-2021-06-17T16-10-50
+        image: gcr.io/istio-testing/build-tools:release-1.10-2021-07-06T17-17-03
         name: ""
         resources:
           limits:
@@ -88,7 +88,7 @@ postsubmits:
         - entrypoint
         - prow/integ-suite-kind.sh
         - doc.test.profile_default
-        image: gcr.io/istio-testing/build-tools:release-1.10-2021-06-17T16-10-50
+        image: gcr.io/istio-testing/build-tools:release-1.10-2021-07-06T17-17-03
         name: ""
         resources:
           limits:
@@ -143,7 +143,7 @@ postsubmits:
         - entrypoint
         - prow/integ-suite-kind.sh
         - doc.test.profile_demo
-        image: gcr.io/istio-testing/build-tools:release-1.10-2021-06-17T16-10-50
+        image: gcr.io/istio-testing/build-tools:release-1.10-2021-07-06T17-17-03
         name: ""
         resources:
           limits:
@@ -198,7 +198,7 @@ postsubmits:
         - entrypoint
         - prow/integ-suite-kind.sh
         - doc.test.profile_none
-        image: gcr.io/istio-testing/build-tools:release-1.10-2021-06-17T16-10-50
+        image: gcr.io/istio-testing/build-tools:release-1.10-2021-07-06T17-17-03
         name: ""
         resources:
           limits:
@@ -255,7 +255,7 @@ postsubmits:
         - --topology
         - MULTICLUSTER
         - doc.test.multicluster
-        image: gcr.io/istio-testing/build-tools:release-1.10-2021-06-17T16-10-50
+        image: gcr.io/istio-testing/build-tools:release-1.10-2021-07-06T17-17-03
         name: ""
         resources:
           limits:
@@ -312,7 +312,7 @@ presubmits:
       - command:
         - make
         - lint
-        image: gcr.io/istio-testing/build-tools:release-1.10-2021-06-17T16-10-50
+        image: gcr.io/istio-testing/build-tools:release-1.10-2021-07-06T17-17-03
         name: ""
         resources:
           limits:
@@ -349,7 +349,7 @@ presubmits:
       - command:
         - make
         - gen-check
-        image: gcr.io/istio-testing/build-tools:release-1.10-2021-06-17T16-10-50
+        image: gcr.io/istio-testing/build-tools:release-1.10-2021-07-06T17-17-03
         name: ""
         resources:
           limits:
@@ -387,7 +387,7 @@ presubmits:
         - entrypoint
         - prow/integ-suite-kind.sh
         - doc.test.profile_default
-        image: gcr.io/istio-testing/build-tools:release-1.10-2021-06-17T16-10-50
+        image: gcr.io/istio-testing/build-tools:release-1.10-2021-07-06T17-17-03
         name: ""
         resources:
           limits:
@@ -443,7 +443,7 @@ presubmits:
         - entrypoint
         - prow/integ-suite-kind.sh
         - doc.test.profile_demo
-        image: gcr.io/istio-testing/build-tools:release-1.10-2021-06-17T16-10-50
+        image: gcr.io/istio-testing/build-tools:release-1.10-2021-07-06T17-17-03
         name: ""
         resources:
           limits:
@@ -499,7 +499,7 @@ presubmits:
         - entrypoint
         - prow/integ-suite-kind.sh
         - doc.test.profile_none
-        image: gcr.io/istio-testing/build-tools:release-1.10-2021-06-17T16-10-50
+        image: gcr.io/istio-testing/build-tools:release-1.10-2021-07-06T17-17-03
         name: ""
         resources:
           limits:
@@ -557,7 +557,7 @@ presubmits:
         - --topology
         - MULTICLUSTER
         - doc.test.multicluster
-        image: gcr.io/istio-testing/build-tools:release-1.10-2021-06-17T16-10-50
+        image: gcr.io/istio-testing/build-tools:release-1.10-2021-07-06T17-17-03
         name: ""
         resources:
           limits:
