--- prow/config/istio-private_jobs/release-builder-1.5.yaml
@@ -1,31 +1,0 @@
-defaults:
-  branches: [release-1.5]
-  repo-allowlist: [release-builder]
-
-
-transforms:
-
-# istio/release-builder release-1.5 test jobs(s) - pre/postsubmit(s)
-- job-type: [presubmit, postsubmit]
-  repo-allowlist: [release-builder]
-  job-allowlist: [lint_release-builder_release-1.5,lint_release-builder_release-1.5_postsubmit,test_release-builder_release-1.5,test_release-builder_release-1.5_postsubmit,gencheck_release-builder_release-1.5,gencheck_release-builder_release-1.5_postsubmit]
-
-# istio/release-builder release-1.5 build warning jobs(s) - presubmit(s)
-- env:
-    PRERELEASE_DOCKER_HUB: gcr.io/istio-prow-build
-    GCS_BUCKET: istio-private-prerelease/prerelease
-  job-type: [presubmit]
-  repo-allowlist: [release-builder]
-  job-allowlist: [build-warning_release-builder_release-1.5]
-
-  # istio/release-builder release-1.5 build jobs(s) - postsubmit(s)
-- env:
-    PRERELEASE_DOCKER_HUB: gcr.io/istio-prow-build
-    GCS_BUCKET: istio-private-prerelease/prerelease
-  labels:
-    preset-enable-ssh: "true"
-    preset-override-envoy: "true"
-    preset-override-deps: release-1.5-release
-  job-type: [postsubmit]
-  repo-allowlist: [release-builder]
-  job-allowlist: [build-release_release-builder_release-1.5_postsubmit]
